migrate DI solution (GSI-435)
Switched to DI based on basic Python constructs like context managers.

Upgrades to pydantic v2.

This PR had to upgrade to the latest ghga-event-schemas which also introduce the s3_endpoint_alias fields.
Since this service is not yet ready to handle them, I just hard-coded the field where required.
Has to be cleaned up in a dedicated PR.

Switches to docker-outside-of-docker for the devcontainer environment.

Bumps version to 1.0.0.